+++ lib/ruby-next/commands/core_ext.rb
@@ -70,9 +70,9 @@ module RubyNext
           .values
           .flatten
           .select do |patch|
-          next if min_version && Gem::Version.new(patch.version) <= min_version
-          next if filter && !filter.match?(patch.name)
-          true
+            next if min_version && Gem::Version.new(patch.version) <= min_version
+            next if filter && !filter.match?(patch.name)
+            true
         end
       end